# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AgriStroom OpenApi Documentation",
        "description": "AgriStroom Swagger OpenApi Documentation",
        "contact": {
            "email": "muhohoweb@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/post/create": {
            "post": {
                "tags": [
                    "create post"
                ],
                "summary": "Adds a new user - with oneOf examples",
                "operationId": "99ac18525eda9ec469e45236535ae5c1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "string"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "integer"
                                            }
                                        ]
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "What is the optimum environment for apple farming"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/PostResource"
                                        },
                                        {
                                            "type": "boolean"
                                        }
                                    ]
                                },
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true
                                        }
                                    },
                                    "bool": {
                                        "summary": "A boolean value.",
                                        "value": false
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/reply/create": {
            "post": {
                "tags": [
                    "create reply"
                ],
                "summary": "Create reply",
                "description": "Creates a reply",
                "operationId": "getProjectsList",
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PostResource": {}
        }
    },
    "tags": [
        {
            "name": "AgriStroom",
            "description": "API Endpoints of AgriStroom"
        }
    ]
}